Reconstruct the setting.rs in core crate
